README.md
Synopsis

This is a simple To Do List App using MEAN stack. 
This will help people add their tasks using the app, and save it as homepage so they remember it.

Motivation

Planning does half the work, but the other options available on the internet are too cluttered with very fancy UI, etc. All we need is 
a simple UI to achieve this. This app will give you minimally confusing UI with basic functionality needed to add/delete/update App.

Installation

You will need to install node.js(along with express, body-parser), mongodb to be able to run this.